Updated git module to latest master from http://github.com/jbarnette/vlad-git
Add support for:
 * fast checkouts (only fetch changes instead of full repo)
 * git submodule support

+++ lib/vlad/git.rb
@@ -1,6 +1,9 @@
 class Vlad::Git
 
-  set :source, Vlad::Git.new
+  # Duh.
+  VERSION = "2.1.0"
+
+  set :source,  Vlad::Git.new
   set :git_cmd, "git"
 
   ##
@@ -12,12 +15,25 @@ class Vlad::Git
     destination = File.join(destination, 'repo')
     revision = 'HEAD' if revision =~ /head/i
 
-    [ "rm -rf #{destination}",
-      "#{git_cmd} clone #{repository} #{destination}",
-      "cd #{destination}",
-      "#{git_cmd} checkout -f -b deployed-#{revision} #{revision}",
-      "cd -"
-    ].join(" && ")
+    if fast_checkout_applicable?(revision, destination)
+      [ "cd #{destination}",
+        "#{git_cmd} checkout -q origin",
+        "#{git_cmd} fetch",
+        "#{git_cmd} reset --hard #{revision}",
+        "#{git_cmd} submodule update --init",
+        "#{git_cmd} branch -f deployed-#{revision} #{revision}",
+        "#{git_cmd} checkout deployed-#{revision}",
+        "cd -"
+      ].join(" && ")
+    else
+      [ "rm -rf #{destination}",
+        "#{git_cmd} clone #{repository} #{destination}",
+        "cd #{destination}",
+        "#{git_cmd} submodule update --init",
+        "#{git_cmd} checkout -f -b deployed-#{revision} #{revision}",
+        "cd -"
+      ].join(" && ")
+    end
   end
 
   ##
@@ -32,6 +48,7 @@ class Vlad::Git
     [ "mkdir -p #{destination}",
       "cd repo",
       "#{git_cmd} archive --format=tar #{revision} | (cd #{destination} && tar xf -)",
+      "#{git_cmd} submodule foreach '#{git_cmd} archive --format=tar $sha1 | (cd #{destination}/$path && tar xf -)'",
       "cd -",
       "cd .."
     ].join(" && ")
@@ -45,5 +62,23 @@ class Vlad::Git
     revision = 'HEAD' if revision =~ /head/i
 
     "`#{git_cmd} rev-parse #{revision}`"
+  end
+
+  private
+
+  # Checks if fast-checkout is applicable
+  def fast_checkout_applicable?(revision, destination)
+    revision = 'HEAD' if revision =~ /head/i
+
+    begin
+      cmd = [ "if cd #{destination}",
+              "#{git_cmd} rev-parse #{revision}",
+              "#{git_cmd} remote -v | grep -q #{repository}",
+              "cd -; then exit 0; else exit 1; fi &>/dev/null" ].join(" && ")
+      run cmd
+      return true
+    rescue Vlad::CommandFailedError
+      return false
+    end
   end
 end
